[MobilePhone] GMCM API interface moved to 'API' folder. Imports updated.
