+++ MobilePhone/ModEntry.cs
@@ -282,6 +282,7 @@ namespace MobilePhone
                 }
             }
         }
+
         public override object GetApi()
         {
             return new MobilePhoneApi();
